// lib/api-types.ts
export enum ArticleStatus {
  Drafted = 'Drafted',
  DraftingInProgress = 'Drafting',
  PendingReview = 'Pending Review',
  ReviewCompleted = 'Review Completed',
  Approved = 'Approved',
  Published = 'Published',
  Failed = 'Failed',
}
// ...
export interface ArticleUpdateRequest {
  title?: string
  content?: string
  status?: ArticleStatus | string
}